backend/requirements.txt
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
Django==2.2.5
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytz==2021.3
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
